requirements.txt
Django==4.1.1
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-storages==1.14.2
django-timezone-field==6.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
nltk==3.8.1
spacy==3.7.1
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.2
transformers==4.34.0